+++ server/odc-core/src/main/java/com/oceanbase/odc/core/shared/constant/LimitMetric.java
@@ -46,7 +46,7 @@ public enum LimitMetric implements Translatable {
     TABLE_NAME_LENGTH,
     WORKSHEET_CHANGE_COUNT,
     WORKSHEET_SAME_LEVEL_COUNT,
-    WORKSHEET_PROJECT_COUNT;
+    WORKSHEET_COUNT_IN_PROJECT;
 
     @Override
     public String code() {
